requirements.txt
pyglet>=1.3
PyGame
PyOpenGL